deploy/requirements.txt
# Taris Bot — Python package dependencies
# Install with: python3 -m pip install --break-system-packages -r requirements.txt
# Last updated: 2026-04-02
#
# Variants:
#   PicoClaw  — Raspberry Pi; STT_PROVIDER=vosk (no faster-whisper)
#   OpenClaw  — Laptop/PC x86_64; STT_PROVIDER=faster_whisper (no vosk required)
#
# Install all:      pip install --break-system-packages -r requirements.txt
# Install picoclaw: pip install --break-system-packages -r requirements.txt --ignore-requires  # skip openclaw section manually
# Install openclaw: pip install --break-system-packages -r requirements.txt

# ── Telegram bot ─────────────────────────────────────────────────────────────
pyTelegramBotAPI

# ── Voice — STT / TTS ────────────────────────────────────────────────────────
# PicoClaw (Raspberry Pi) — Vosk offline STT
vosk                   # offline Russian/German STT
sounddevice            # audio capture via PortAudio
webrtcvad              # optional — VAD pre-filter; graceful fallback if absent

# OpenClaw (x86_64 laptop/PC) — faster-whisper STT (better WER than Vosk)
# Requires: PIPER_BIN and PIPER_MODEL set in bot.env (NOT hardcoded path)
# On Debian Bookworm: install with --break-system-packages
faster-whisper         # CTranslate2-based Whisper — STT_PROVIDER=faster_whisper
scipy                  # audio resampling when sample_rate != 16000 (used by faster-whisper path)

# ── Gmail digest — Google API ─────────────────────────────────────────────────
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# ── Web UI (FastAPI) ──────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
jinja2
bcrypt
PyJWT
python-multipart
requests               # Google OAuth token refresh + HTTP calls (bot_web.py)

# ── Embeddings ────────────────────────────────────────────────────────────────
# ONNX Runtime-based; ARM-safe (replaces sentence_transformers/torch which OOMs on Pi 3)
fastembed              # pulls in onnxruntime-cpu automatically; model: all-MiniLM-L6-v2 384-dim

# ── Storage — PostgreSQL + pgvector ──────────────────────────────────────────
# Optional: required only when STORE_BACKEND=postgres (OpenClaw default on TariStation2)
psycopg[binary]
psycopg-pool
pgvector
cryptography           # TLS for psycopg connections

# ── Screen DSL ───────────────────────────────────────────────────────────────
pyyaml                 # YAML screen definitions (src/screens/)
jsonschema             # optional — YAML/JSON screen file validation

# ── LLM providers ────────────────────────────────────────────────────────────
# Used when LLM_PROVIDER=openai (OpenAI API or OpenRouter-compatible API)
openai                 # required for LLM_PROVIDER=openai (gpt-4o-mini etc.)
                       # Ollama runs as a local service — no Python package needed

# ── Documents / RAG ──────────────────────────────────────────────────────────
sentence-transformers  # all-MiniLM-L6-v2 embeddings (~90 MB)
pdfminer.six           # PDF text extraction
python-docx            # .docx text extraction
sqlite-vec             # SQLite vector search extension (RAG / KNN embeddings)

# ── Remote KB — MCP client (N8N MCP Server Trigger, Phase 3+) ────────────────
# Required when REMOTE_KB_ENABLED=1. Safe to install on all targets (pure Python).
mcp>=1.2.0             # MCP protocol client (SSE transport)
httpx>=0.27.0          # async HTTP used by mcp SSE client
httpx-sse>=0.4.0       # SSE streaming for httpx
